OMAP4: DPLL cascading: mute warnings during DPLL cascading
1) PM: mute warn from min_bus_tput constraint

When the omap_device_set_rate function is called from
omap_pm_set_min_bus_tput and fails, a warn is printed.  This behavior is
normal and expected in DPLL cascading mode, so the warn has been
converted to pr_debug to minimize console noise.

2)clock: mute WARN when searching for fieldval

This warn is hit every time when entering DPLL cascading.  This is due
to a workaround implemented in DPLL cascading code which must circumvent the
clock framework to change l4_wkup_iclk's parent from SYS_CK to
DPLL_ABE's LP output, and back again.
During low power MP3 playback it is observed that this warn can disrupt
playback by causing a rapid exit/re-entry of the DPLL cascading mode.

[Andrii Tseglytskyi <andrii.tseglytskyi@ti.com>:
Ported and adapted to p-android-omap-3.0]

Change-Id: Id040a4710303845ebc9eeca81c1c06d49f922230
Signed-off-by: Mike Turquette <mturquette@ti.com>
Signed-off-by: Andrii Tseglytskyi <andrii.tseglytskyi@ti.com>