Revert "configs: blaze/panda: zero out remoteproc dynamic carveout (dsp)"
This reverts commit 9e69eab645d01b7b6c8d6d9eda64eaa210236536.

DSP currently requires 4MB of memory to be set-aside to meet its
firmware requirements. This memory can either be dynamically
reserved or statically reserved during board initialization.
The above commit zeroed out the dynamic configuration sizes as
the static reservation was used previously. DSP does not have
a mandatory requirement to use static reservation, hence the
above commit is being reverted in preparation for switching to
a dynamic reservation.

Change-Id: I8bf374b7865025f2e6b3aff7c5e3308a136a87ea
Signed-off-by: Jesse Villarreal <jesse.villarreal@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>