gcbv: improve check for fill condition
When a usecase generates a fill with a 1x1 destination
we were hitting the blt rather than fill path. This change
improves the check for fill to validate whether there is
a virtual address which the fill code needs anyway as it
directly accessing the memory provided by the src1 buffer
descriptor.

Change-Id: Iafae2dcb8a9b7020d07c6ac7f46a950efb6e4996
Signed-off-by: Tony Lofthouse <a0741364@ti.com>